auth_service/requirements.txt
Flask==2.1.1
Authlib==0.15.5
mysql-connector-python==8.0.31
werkzeug==2.0.3
flask-cors==3.0.10
